streamlined

+++ graph.h
@@ -30,6 +30,6 @@ int countInLinks(Graph g, GraphPage url);
 int isInLink(Graph g, GraphPage u, GraphPage v);
 double inLinkPopularity(Graph, GraphPage, GraphPage);
 double outLinkPopularity(Graph, GraphPage, GraphPage);
-int countOutLinks(Graph, int);
+double countOutLinks(Graph, int);
 
 #endif
